complete run_star move to tasks module
reorganize run_star dependencies
use main function in ensembl_anno
consolidate single message logging messages
use pathlib Path objects for paths
